meteo/requirements.txt
django==4.2.14
openmeteo-requests==1.2.0
requests-cache==1.2.1
retry-requests==2.0.0
numpy==1.24.4
pandas==2.0.3
django-pandas==0.6.7
